Fix conversion warnings in ParticleShape
There were are couple conversion warnings from size_t to double, as well
as a constructor warning that came up when building with Visual Studio.

+++ SurgSim/Math/ParticlesShape.cpp
@@ -69,7 +69,7 @@ void ParticlesShape::setRadius(double radius)
 
 bool ParticlesShape::doUpdate()
 {
-	const size_t numParticles = getVertices().size();
+	const double numParticles = static_cast<double>(getVertices().size());
 	const Vector3d radius = Vector3d::Constant(m_radius);
 
 	std::list<DataStructures::AabbTreeData::Item> items;
